web-application/backend/requirements.txt
aniso8601==8.1.0
cachelib==0.1.1
click==7.1.2
cycler==0.10.0
Flask==1.1.2
Flask-Cors==3.0.10
Flask-RESTful==0.3.8
Flask-Session==0.3.2
gensim==4.0.0b0
gunicorn==20.0.4
itsdangerous==1.1.0
Jinja2==2.11.2
joblib==1.0.0
kiwisolver==1.3.1
MarkupSafe==1.1.1
marshmallow==3.10.0
matplotlib==3.3.3
numpy==1.19.5
Pillow==8.1.0
pyparsing==2.4.7
python-dateutil==2.8.1
python-dotenv==0.15.0
pytz==2020.5
scikit-learn==0.24.1
scipy==1.6.0
six==1.15.0
smart-open==4.1.2
threadpoolctl==2.1.0
webargs==7.0.1
Werkzeug==1.0.1
